ZoweUSSNode.unit.test.ts refactoring
-Moved all tests for ZoweUSSNode functions into __tests__/__unit__/uss/ZoweUSSNode.unit.test.ts
-Created mock generator file for uss mocks (__mocks__/generators/uss.ts)
