+++ __tests__/__unit__/extension.unit.test.ts
@@ -2427,117 +2427,6 @@ describe("Extension Unit Tests", () => {
         expect(spyopenPS).toHaveBeenCalled();
     });
 
-    describe("refresh USS checking", () => {
-        const isDirtyInEditor = jest.fn();
-        const openedDocumentInstance = jest.fn();
-
-        const setMocksForNode = (node: ZoweUSSNode) => {
-            Object.defineProperty(node, "isDirtyInEditor", {
-                get: isDirtyInEditor
-            });
-            Object.defineProperty(node, "openedDocumentInstance", {
-                get: openedDocumentInstance
-            });
-
-            node.contextValue = globals.USS_SESSION_CONTEXT;
-            node.fullPath = "/u/myuser";
-        };
-        const resetMocks = () => {
-            showErrorMessage.mockReset();
-            showTextDocument.mockReset();
-            ussFile.mockReset();
-            executeCommand.mockReset();
-            isDirtyInEditor.mockReset();
-            openedDocumentInstance.mockReset();
-        };
-
-        it("refreshUSS works correctly for dirty file state, when user didn't cancel file save", async () => {
-            const node = new ZoweUSSNode("test-node", vscode.TreeItemCollapsibleState.None, ussNode, null, "/");
-
-            resetMocks();
-            setMocksForNode(node);
-
-            const response: zowe.IZosFilesResponse = {
-                success: true,
-                commandResponse: null,
-                apiResponse: {
-                    etag: "132"
-                }
-            };
-            ussFile.mockResolvedValue(response);
-            isDirtyInEditor.mockReturnValueOnce(true);
-            isDirtyInEditor.mockReturnValueOnce(false);
-            await node.refreshUSS();
-
-            expect(ussFile.mock.calls.length).toBe(1);
-            expect(showTextDocument.mock.calls.length).toBe(2);
-            expect(executeCommand.mock.calls.length).toBe(2);
-            expect(node.downloaded).toBe(true);
-        });
-        it("refreshUSS works correctly for dirty file state, when user cancelled file save", async () => {
-            const node = new ZoweUSSNode("test-node", vscode.TreeItemCollapsibleState.None, ussNode, null, "/");
-
-            resetMocks();
-            setMocksForNode(node);
-
-            const response: zowe.IZosFilesResponse = {
-                success: true,
-                commandResponse: null,
-                apiResponse: {
-                    etag: "132"
-                }
-            };
-            ussFile.mockResolvedValueOnce(response);
-            isDirtyInEditor.mockReturnValueOnce(true);
-            isDirtyInEditor.mockReturnValueOnce(true);
-            await node.refreshUSS();
-
-            expect(ussFile.mock.calls.length).toBe(0);
-            expect(showTextDocument.mock.calls.length).toBe(1);
-            expect(executeCommand.mock.calls.length).toBe(1);
-            expect(node.downloaded).toBe(false);
-        });
-        it("refreshUSS works correctly for not dirty file state", async () => {
-            const node = new ZoweUSSNode("test-node", vscode.TreeItemCollapsibleState.None, ussNode, null, "/");
-
-            resetMocks();
-            setMocksForNode(node);
-
-            const response: zowe.IZosFilesResponse = {
-                success: true,
-                commandResponse: null,
-                apiResponse: {
-                    etag: "132"
-                }
-            };
-            ussFile.mockResolvedValueOnce(response);
-            isDirtyInEditor.mockReturnValueOnce(false);
-            isDirtyInEditor.mockReturnValueOnce(false);
-            await node.refreshUSS();
-
-            expect(ussFile.mock.calls.length).toBe(1);
-            expect(showTextDocument.mock.calls.length).toBe(0);
-            expect(executeCommand.mock.calls.length).toBe(1);
-            expect(node.downloaded).toBe(true);
-        });
-        it("refreshUSS works correctly with exception thrown in process", async () => {
-            const node = new ZoweUSSNode("test-node", vscode.TreeItemCollapsibleState.None, ussNode, null, "/");
-
-            resetMocks();
-            setMocksForNode(node);
-
-            ussFile.mockRejectedValueOnce(Error(""));
-            isDirtyInEditor.mockReturnValueOnce(true);
-            isDirtyInEditor.mockReturnValueOnce(false);
-            await node.refreshUSS();
-
-            expect(ussFile.mock.calls.length).toBe(1);
-            expect(showTextDocument.mock.calls.length).toBe(1);
-            expect(executeCommand.mock.calls.length).toBe(1);
-            expect(node.downloaded).toBe(false);
-        });
-    });
-
     describe("Add searchForLoadedItems Tests", () => {
         it("Testing that filterTreeByString returns the correct array", async () => {
             const qpItems = [
@@ -3079,386 +2968,6 @@ describe("Extension Unit Tests", () => {
             expect(extension.addZoweSession).toHaveBeenCalled();
 
         });
-    });
-
-  // TODO Node tests
-    it("Testing that open is executed successfully", async () => {
-        ussFile.mockReset();
-        openTextDocument.mockReset();
-        showTextDocument.mockReset();
-        showErrorMessage.mockReset();
-        existsSync.mockReset();
-        withProgress.mockReset();
-
-        Object.defineProperty(Profiles, "getInstance", {
-            value: jest.fn(() => {
-                return {
-                    allProfiles: [{name: "firstName"}, {name: "secondName"}],
-                    defaultProfile: {name: "firstName"},
-                    getDefaultProfile: mockLoadNamedProfile,
-                    promptCredentials: jest.fn(()=> {
-                        return ["fake", "fake", "fake"];
-                    }),
-                    loadNamedProfile: mockLoadNamedProfile,
-                    usesSecurity: true,
-                    validProfile: ValidProfileEnum.VALID,
-                    checkCurrentProfile: jest.fn(),
-                    getProfiles: jest.fn(() => {
-                        return [{name: profileOne.name, profile: profileOne}, {name: profileOne.name, profile: profileOne}];
-                    }),
-                    refresh: jest.fn(),
-                };
-            })
-        });
-
-        const node = new ZoweUSSNode("node", vscode.TreeItemCollapsibleState.None, ussNode, session, "/", false, profileOne.name);
-        const parent = new ZoweUSSNode("parent", vscode.TreeItemCollapsibleState.Collapsed, ussNode, null, "/", false, profileOne.name);
-        const child = new ZoweUSSNode("child", vscode.TreeItemCollapsibleState.None, parent, null, "/parent", false, profileOne.name);
-
-        const isBinSpy = jest.spyOn(ussApi, "isFileTagBinOrAscii");
-        existsSync.mockReturnValue(null);
-        openTextDocument.mockResolvedValueOnce("test.doc");
-
-        ussFile.mockReturnValueOnce(fileResponse);
-        withProgress.mockReturnValue(fileResponse);
-
-        await node.openUSS(false, true, testUSSTree);
-
-        expect(existsSync.mock.calls.length).toBe(1);
-        expect(existsSync.mock.calls[0][0]).toBe(path.join(globals.USS_DIR, "/" + node.mProfileName + "/", node.fullPath));
-        expect(isFileTagBinOrAscii.mock.calls.length).toBe(1);
-        expect(isFileTagBinOrAscii.mock.calls[0][0]).toBe(session);
-        expect(isFileTagBinOrAscii.mock.calls[0][1]).toBe(node.fullPath);
-        expect(withProgress).toBeCalledWith(
-            {
-                location: vscode.ProgressLocation.Notification,
-                title: "Opening USS file..."
-            }, expect.any(Function)
-        );
-        withProgress(downloadUSSFile);
-        expect(withProgress).toBeCalledWith(downloadUSSFile);
-        expect(openTextDocument.mock.calls.length).toBe(1);
-        expect(openTextDocument.mock.calls[0][0]).toBe(node.getUSSDocumentFilePath());
-        expect(showTextDocument.mock.calls.length).toBe(1);
-        expect(showTextDocument.mock.calls[0][0]).toBe("test.doc");
-
-        openTextDocument.mockResolvedValueOnce("test.doc");
-        const node2 = new ZoweUSSNode("usstest", vscode.TreeItemCollapsibleState.None, ussNode, null, null);
-
-        await node2.openUSS(false, true, testUSSTree);
-
-        ussFile.mockReset();
-        openTextDocument.mockReset();
-        showTextDocument.mockReset();
-        existsSync.mockReset();
-
-        existsSync.mockReturnValue("exists");
-        showTextDocument.mockRejectedValueOnce(Error("testError"));
-
-        try {
-            await child.openUSS(false, true, testUSSTree);
-        } catch (err) {
-            // do nothing
-        }
-
-        expect(ussFile.mock.calls.length).toBe(0);
-        expect(openTextDocument.mock.calls.length).toBe(1);
-        expect(openTextDocument.mock.calls[0][0]).toBe(child.getUSSDocumentFilePath());
-        expect(showTextDocument.mock.calls.length).toBe(1);
-        expect(showErrorMessage.mock.calls.length).toBe(1);
-        expect(showErrorMessage.mock.calls[0][0]).toBe("testError Error: testError");
-
-        const child2 = new ZoweUSSNode("child", vscode.TreeItemCollapsibleState.None, node2, null, null);
-        try {
-            await child2.openUSS(false, true, testUSSTree);
-        } catch (err) {
-            // do nothing
-        }
-
-        ussFile.mockReset();
-        openTextDocument.mockReset();
-        showTextDocument.mockReset();
-        existsSync.mockReset();
-        showErrorMessage.mockReset();
-
-        const badparent = new ZoweUSSNode("parent", vscode.TreeItemCollapsibleState.Collapsed, ussNode, null, null);
-        badparent.contextValue = "turnip";
-        const brat = new ZoweUSSNode("brat", vscode.TreeItemCollapsibleState.None, badparent, null, null);
-        try {
-            await brat.openUSS(false, true, testUSSTree);
-// tslint:disable-next-line: no-empty
-        } catch (err) {
-        }
-        expect(ussFile.mock.calls.length).toBe(0);
-        expect(showErrorMessage.mock.calls.length).toBe(2);
-        expect(showErrorMessage.mock.calls[0][0]).toBe("open() called from invalid node.");
-        expect(showErrorMessage.mock.calls[1][0]).toBe("open() called from invalid node. Error: open() called from invalid node.");
-    });
-
-    // TODO Node tests
-    it("Tests that openUSS executes successfully with favored files", async () => {
-        ussFile.mockReset();
-        openTextDocument.mockReset();
-        showTextDocument.mockReset();
-
-        ussFile.mockReturnValueOnce(fileResponse);
-
-        openTextDocument.mockResolvedValueOnce("test.doc");
-
-        Object.defineProperty(Profiles, "getInstance", {
-            value: jest.fn(() => {
-                return {
-                    allProfiles: [{name: "firstName"}, {name: "secondName"}],
-                    defaultProfile: {name: "firstName"},
-                    getDefaultProfile: mockLoadNamedProfile,
-                    promptCredentials: jest.fn(()=> {
-                        return ["fake", "fake", "fake"];
-                    }),
-                    loadNamedProfile: mockLoadNamedProfile,
-                    usesSecurity: true,
-                    validProfile: ValidProfileEnum.VALID,
-                    checkCurrentProfile: jest.fn(),
-                    getProfiles: jest.fn(() => {
-                        return [{name: profileOne.name, profile: profileOne}, {name: profileOne.name, profile: profileOne}];
-                    }),
-                    refresh: jest.fn(),
-                };
-            })
-        });
-
-        // Set up mock favorite session
-        const favoriteSession = new ZoweUSSNode("Favorites", vscode.TreeItemCollapsibleState.Collapsed, null, session, null, false, profileOne.name);
-        favoriteSession.contextValue = globals.FAVORITE_CONTEXT;
-
-        // Set up favorited nodes (directly under Favorites)
-        const favoriteFile = new ZoweUSSNode("favFile", vscode.TreeItemCollapsibleState.None, favoriteSession, session, "/", false, profileOne.name);
-        favoriteFile.contextValue = globals.DS_TEXT_FILE_CONTEXT + globals.FAV_SUFFIX;
-        const favoriteParent = new ZoweUSSNode("favParent", vscode.TreeItemCollapsibleState.Collapsed, favoriteSession, null, "/",
-            false, profileOne.name);
-        favoriteParent.contextValue = globals.USS_DIR_CONTEXT + globals.FAV_SUFFIX;
-        // Set up child of favoriteDir - make sure we can open the child of a favored directory
-        const child = new ZoweUSSNode("favChild", vscode.TreeItemCollapsibleState.Collapsed, favoriteParent, null, "/favDir", false, profileOne.name);
-        child.contextValue = globals.DS_TEXT_FILE_CONTEXT;
-
-        // For each node, make sure that code below the log.debug statement is execute
-        await favoriteFile.openUSS(false, true, testUSSTree);
-        expect(showTextDocument.mock.calls.length).toBe(1);
-        showTextDocument.mockReset();
-        ussFile.mockReturnValueOnce(fileResponse);
-        await child.openUSS(false, true, testUSSTree);
-        expect(showTextDocument.mock.calls.length).toBe(1);
-        showTextDocument.mockReset();
-    });
-
-    // TODO Node tests
-    it("Testing that open is executed successfully when chtag says binary", async () => {
-        ussFile.mockReset();
-        openTextDocument.mockReset();
-        showTextDocument.mockReset();
-        showErrorMessage.mockReset();
-        existsSync.mockReset();
-        withProgress.mockReset();
-
-        Object.defineProperty(Profiles, "getInstance", {
-            value: jest.fn(() => {
-                return {
-                    allProfiles: [{name: "firstName"}, {name: "secondName"}],
-                    defaultProfile: {name: "firstName"},
-                    getDefaultProfile: mockLoadNamedProfile,
-                    promptCredentials: jest.fn(()=> {
-                        return ["fake", "fake", "fake"];
-                    }),
-                    loadNamedProfile: mockLoadNamedProfile,
-                    usesSecurity: true,
-                    validProfile: ValidProfileEnum.VALID,
-                    checkCurrentProfile: jest.fn(),
-                    getProfiles: jest.fn(() => {
-                        return [{name: profileOne.name, profile: profileOne}, {name: profileOne.name, profile: profileOne}];
-                    }),
-                    refresh: jest.fn(),
-                };
-            })
-        });
-
-        const node = new ZoweUSSNode("node", vscode.TreeItemCollapsibleState.None, ussNode, session, "/", false, ussNode.getProfileName());
-        const parent = new ZoweUSSNode("parent", vscode.TreeItemCollapsibleState.Collapsed, ussNode, null, "/");
-        const child = new ZoweUSSNode("child", vscode.TreeItemCollapsibleState.None, parent, null, "/parent");
-
-        isFileTagBinOrAscii.mockReturnValue(true);
-        existsSync.mockReturnValue(null);
-        openTextDocument.mockResolvedValueOnce("test.doc");
-
-        withProgress.mockReturnValue(fileResponse);
-
-        await node.openUSS(false, true, testUSSTree);
-
-        expect(existsSync.mock.calls.length).toBe(1);
-        expect(existsSync.mock.calls[0][0]).toBe(path.join(globals.USS_DIR, "/" + node.getProfileName() + "/", node.fullPath));
-        expect(withProgress).toBeCalledWith(
-            {
-                location: vscode.ProgressLocation.Notification,
-                title: "Opening USS file..."
-            }, expect.any(Function)
-        );
-        withProgress(downloadUSSFile);
-        expect(withProgress).toBeCalledWith(downloadUSSFile);
-        expect(openTextDocument.mock.calls.length).toBe(1);
-        expect(openTextDocument.mock.calls[0][0]).toBe(node.getUSSDocumentFilePath());
-        expect(showTextDocument.mock.calls.length).toBe(1);
-        expect(showTextDocument.mock.calls[0][0]).toBe("test.doc");
-    });
-
-        // TODO Node tests
-    it("Testing that that openUSS credentials prompt is executed successfully", async () => {
-        showQuickPick.mockReset();
-        showInputBox.mockReset();
-        showTextDocument.mockReset();
-        openTextDocument.mockReset();
-        const sessionwocred = new imperative.Session({
-            user: "",
-            password: "",
-            hostname: "fake",
-            port: 443,
-            protocol: "https",
-            type: "basic",
-        });
-
-        ussFile.mockReturnValueOnce(fileResponse);
-        const dsNode = new ZoweUSSNode("testSess", vscode.TreeItemCollapsibleState.Expanded, ussNode, sessionwocred, null);
-        dsNode.contextValue = globals.USS_SESSION_CONTEXT;
-        Object.defineProperty(Profiles, "getInstance", {
-            value: jest.fn(() => {
-                return {
-                    allProfiles: [{name: "firstName", profile: {user:undefined, password: undefined}}, {name: "secondName"}],
-                    defaultProfile: {name: "firstName"},
-                    promptCredentials: jest.fn(()=> {
-                        return ["fake", "fake", "fake"];
-                    }),
-                    loadNamedProfile: mockLoadNamedProfile,
-                    validProfile: ValidProfileEnum.VALID,
-                    checkCurrentProfile: jest.fn(),
-                };
-            })
-        });
-
-        showInputBox.mockReturnValueOnce("fake");
-        showInputBox.mockReturnValueOnce("fake");
-
-        await dsNode.openUSS(false, true, testUSSTree);
-        expect(openTextDocument.mock.calls.length).toBe(1);
-        expect(showTextDocument.mock.calls.length).toBe(1);
-    });
-
-    it("Testing that that openUSS credentials prompt works with favorites", async () => {
-        showTextDocument.mockReset();
-        openTextDocument.mockReset();
-        showQuickPick.mockReset();
-        showInputBox.mockReset();
-        const sessionwocred = new imperative.Session({
-            user: "",
-            password: "",
-            hostname: "fake",
-            port: 443,
-            protocol: "https",
-            type: "basic",
-        });
-        const dsNode = new ZoweUSSNode("testSess", vscode.TreeItemCollapsibleState.Expanded, ussNode, sessionwocred, null);
-        dsNode.contextValue = globals.USS_DIR_CONTEXT + globals.FAV_SUFFIX;
-        Object.defineProperty(Profiles, "getInstance", {
-            value: jest.fn(() => {
-                return {
-                    allProfiles: [{name: "firstName", profile: {user:undefined, password: undefined}}, {name: "secondName"}],
-                    defaultProfile: {name: "firstName"},
-                    promptCredentials: jest.fn(()=> {
-                        return ["fake", "fake", "fake"];
-                    }),
-                    loadNamedProfile: mockLoadNamedProfile,
-                    validProfile: ValidProfileEnum.VALID,
-                    checkCurrentProfile: jest.fn(),
-                };
-            })
-        });
-
-        showInputBox.mockReturnValueOnce("fake");
-        showInputBox.mockReturnValueOnce("fake");
-
-        ussFile.mockReturnValueOnce(fileResponse);
-
-        await dsNode.openUSS(false, true, testUSSTree);
-        expect(openTextDocument.mock.calls.length).toBe(1);
-        expect(showTextDocument.mock.calls.length).toBe(1);
-    });
-
-    it("Testing that that openUSS credentials prompt with favorites ends in error", async () => {
-        showTextDocument.mockReset();
-        openTextDocument.mockReset();
-        showQuickPick.mockReset();
-        showInputBox.mockReset();
-        const sessionwocred = new imperative.Session({
-            user: "",
-            password: "",
-            hostname: "fake",
-            port: 443,
-            protocol: "https",
-            type: "basic",
-        });
-        const dsNode = new ZoweUSSNode("testSess", vscode.TreeItemCollapsibleState.Expanded, ussNode, sessionwocred, null);
-        dsNode.contextValue = globals.USS_DIR_CONTEXT + globals.FAV_SUFFIX;
-        Object.defineProperty(Profiles, "getInstance", {
-            value: jest.fn(() => {
-                return {
-                    allProfiles: [{name: "firstName", profile: {user:undefined, password: undefined}}, {name: "secondName"}],
-                    defaultProfile: {name: "firstName"},
-                    validProfile: ValidProfileEnum.INVALID,
-                    checkCurrentProfile: jest.fn(),
-                    promptCredentials: jest.fn(()=> {
-                        return [undefined, undefined, undefined];
-                    }),
-                    loadNamedProfile: mockLoadNamedProfile
-                };
-            })
-        });
-
-        showInputBox.mockReturnValueOnce("fake");
-        showInputBox.mockReturnValueOnce("fake");
-        const spyopenUSS = jest.spyOn(dsNode, "openUSS");
-        await dsNode.openUSS(false, true, testUSSTree);
-        expect(Profiles.getInstance().validProfile).toBe(ValidProfileEnum.INVALID);
-    });
-
-    it("Testing that that openUSS credentials prompt ends in error", async () => {
-        showTextDocument.mockReset();
-        openTextDocument.mockReset();
-        showQuickPick.mockReset();
-        showInputBox.mockReset();
-        const sessionwocred = new imperative.Session({
-            user: "",
-            password: "",
-            hostname: "fake",
-            port: 443,
-            protocol: "https",
-            type: "basic",
-        });
-        const dsNode = new ZoweUSSNode("testSess", vscode.TreeItemCollapsibleState.Expanded, ussNode, sessionwocred, null);
-        dsNode.contextValue = globals.USS_SESSION_CONTEXT;
-        Object.defineProperty(Profiles, "getInstance", {
-            value: jest.fn(() => {
-                return {
-                    allProfiles: [{name: "firstName", profile: {user:undefined, password: undefined}}, {name: "secondName"}],
-                    defaultProfile: {name: "firstName"},
-                    validProfile: ValidProfileEnum.INVALID,
-                    checkCurrentProfile: jest.fn(),
-                    loadNamedProfile: mockLoadNamedProfile
-                };
-            })
-        });
-
-        await dsNode.openUSS(false, true, testUSSTree);
-        expect(Profiles.getInstance().validProfile).toBe(ValidProfileEnum.INVALID);
-        showQuickPick.mockReset();
-        showInputBox.mockReset();
-        showInformationMessage.mockReset();
-        showErrorMessage.mockReset();
     });
 
     it("Testing that saveUSSFile is executed successfully", async () => {
